security update

+++ src/renderer/src/functions/requests.tsx
@@ -1,31 +1,52 @@
 import type { ProjectFormSubmitType, ProjectType } from "@renderer/types"
 import { handleError } from "./handleError"
-import { ResultFromBackendType} from "@renderer/subComponents/ProjectForm"
-export const fetchRequest =async(user_id:string):Promise<ProjectType[] | []>=>{
-    try{
-         const res = await fetch(`https://my-next-dev-project.onrender.com/database/project-ideas?user=${encodeURIComponent(user_id)}`)
+import { ResultFromBackendType} from "@renderer/components/FormComponents/ProjectForm"
+import { CategoriesTypeObjArr } from "@renderer/types"
+import { firebaseStore } from "@renderer/store/projectStore"
+import {type User } from "firebase/auth"
+type UserTokenObjType = {
+    user: User,
+    token:string
+}
+const getFirebaseAuthToken = async():Promise<UserTokenObjType| null>=>{
+  const user:User|null = firebaseStore.getState().user
+    if (!user) {
+        console.warn("No logged-in user");
+        return null;
+        }
+    const token = await user.getIdToken()
+    return {user,token}
+}
+export const fetchRequest =async():Promise<ProjectType[] | []>=>{
+    try{         
+       const auth = await getFirebaseAuthToken()
+        if (!auth) return [];
+         const res = await fetch(`https://my-next-dev-project.onrender.com/database/project-ideas`,{
+            headers:{
+                "Authorization": `Bearer ${auth.token}`
+            }
+         })
      if (!res.ok) {
       console.error("Fetch failed with status", res.status);
       return [];
     }
-const data = await res.json();
+    const data = await res.json();
     console.log(data, "data from backend");
-
     if (!data.results) return []; // user may have 0 projects
     return data.results;
-   
     }catch(err){
         handleError(err,"fetchRequest")
-        return []
-        
+        return []     
     }
 }
 export const postRequest = async(body: Array<ProjectFormSubmitType>):Promise<ResultFromBackendType |null>=>{
-    
+     const auth = await getFirebaseAuthToken()
+        if (!auth) return null;
     try{
         const response = await fetch(`https://my-next-dev-project.onrender.com/database/write-new-project`,{
     method:"POST",
     headers:{
+         "Authorization": `Bearer ${auth.token}`,
         "Content-Type":"application/json"
     },
     body:JSON.stringify(body)
@@ -40,13 +61,17 @@ export const postRequest = async(body: Array<ProjectFormSubmitType>):Promise<Res
     }
 }
 
-export const patchRequest = async(id:string,user_id:string, body:Record<string,any>):Promise<ResultFromBackendType |null>=>{
-    try{
+export const patchRequest = async(id:string, body:Record<string,any>):Promise<ResultFromBackendType |null>=>{
+    try{ 
+     const auth = await getFirebaseAuthToken()
+        if (!auth) return null;
         console.log(body, "body")
-    const response = await fetch(`https://my-next-dev-project.onrender.com/database/project-ideas/${id}/edit?user=${encodeURIComponent(user_id)}`,{
+    const response = await fetch(`https://my-next-dev-project.onrender.com/database/project-ideas/${id}/edit`,{
         method:"PATCH",
         headers:{
-        "Content-Type":"application/json"
+        "Content-Type":"application/json",
+        "Authorization": `Bearer ${auth.token}`,
+        
     },
         body:JSON.stringify(body)
 })
@@ -58,10 +83,15 @@ export const patchRequest = async(id:string,user_id:string, body:Record<string,a
         return null
     }
 }
-export const deleteRequest =async(id:string, user_id:string):Promise<void |null>=>{
+export const deleteRequest =async(id:string):Promise<void |null>=>{
     try{
-        const res = await fetch(`https://my-next-dev-project.onrender.com/database/project-ideas/${id}?user=${encodeURIComponent(user_id)}`,
-            {method: "DELETE"
+        const auth = await getFirebaseAuthToken()
+        if (!auth) return null;
+        const res = await fetch(`https://my-next-dev-project.onrender.com/database/project-ideas/${id}}`,
+            {method: "DELETE",
+                headers:{
+                "Authorization": `Bearer ${auth.token}`
+            }
         })
         const data = await res.json()
          console.log("Project was deleted")
@@ -71,17 +101,30 @@ export const deleteRequest =async(id:string, user_id:string):Promise<void |null>
         return null
     }
 }
-
-export const fetchFilteredRequest =async():Promise<ProjectType[] | null>=>{
+type ProjectFilterSubmitType ={
+   categories: CategoriesTypeObjArr,
+   completed:boolean,
+   
+}
+export const fetchFilteredRequest =async(filtersApplied:ProjectFilterSubmitType):Promise<ResultFromBackendType |[]>=>{
     try{
-         const res = await fetch(`https://my-next-dev-project.onrender.com/database/project-ideas`)
+     const auth = await getFirebaseAuthToken()
+        if (!auth) return [];
+          const res = await fetch(`https://my-next-dev-project.onrender.com/database/project-ideas/filter`,{
+        method:"GET",
+        headers:{
+        "Content-Type":"application/json",
+        "Authorization": `Bearer ${auth.token}`,
+    },
+        body:JSON.stringify(filtersApplied)
+})
     const data = await res.json()
     console.log(data.results, "data")
      if(!data)console.log("something went wrong", data)
     return data.results
     }catch(err){
         handleError(err,"fetchRequest")
-        return null
+        return []
         
     }
 }
